requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose
passlib[bcrypt]
cryptography
celery
redis
requests
python-dotenv
